\name{TopoSUB_restrOutput}
\alias{TopoSUB_restrOutput}

\title{
Restructure TopoSUB - GEOtop output
}
\description{
GEOtop point output retrieved from the coupling with the topoSUB tool is in oder to be used by the topoSUB postprocess functions, e.g. \emph{TopoSUB_mapping}}
\usage{
TopoSUB_restrOutput(wpath, keys, Nclust)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wpath}{
GEOtop simulation path
}
  \item{keys}{
keywords, for further information refer to \pkg{geotopbricks}
}
  \item{Nclust}{
Number of cluster centroids simulated
}
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
